Judge catch juice stream objects at whole-millisecond times like stable

+++ osu.Game.Rulesets.Catch/Objects/JuiceStream.cs
@@ -119,7 +119,8 @@ namespace osu.Game.Rulesets.Catch.Objects
                         AddNested(new Droplet
                         {
                             Samples = dropletSamples,
-                            StartTime = e.Time,
+                            // stable judged these at whole milliseconds.
+                            StartTime = (int)e.Time,
                             X = EffectiveX + Path.PositionAt(e.PathProgress).X,
                         });
                         break;
@@ -130,7 +131,7 @@ namespace osu.Game.Rulesets.Catch.Objects
                         AddNested(new Fruit
                         {
                             Samples = this.GetNodeSamples(nodeIndex++),
-                            StartTime = e.Time,
+                            StartTime = (int)e.Time,
                             X = EffectiveX + Path.PositionAt(e.PathProgress).X,
                         });
                         break;
